♻️ [refactor]: #452 Phase 8 メイン: lexer_read_integer.rs を分離
read_integer の Phase 8-A〜8-E + 8-G〜8-H の 45 テストを
`src/lexer/tests/lexer_read_integer.rs` に分離。Phase 8-F の 4 テストは
次コミットで lexer_read_integer_then_read_real.rs に分離するため
tests.rs に一時的に残す。

- 293 件グリーン維持
- 短縮テスト関数名集合不変 / 重複 0
- lexer_read_integer.rs: 406 行 (< 800)

Refs #452
